backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.5
pydantic-settings==2.7.1
langchain==0.2.16
langchain-groq==0.1.9
langchain-google-genai==1.0.10
langchain-core==0.2.39
langchain-community==0.2.17
prisma==0.15.0
httpx==0.28.1
asyncpg==0.30.0
psycopg2-binary==2.9.10
